+++ src/cli/options.ts
@@ -319,6 +319,11 @@ export const optionMap = {
     alias: '--referrer <referrer>',
     description: 'The referrer for ArNS purchase tracking',
   },
+  reassignAffiliatedNames: {
+    alias: '--reassign-affiliated-names',
+    description: 'Reassign all affiliated names to the new process',
+    type: 'boolean',
+  },
 };
 
 export const walletOptions = [
@@ -466,6 +471,13 @@ export const setAntBaseNameOptions = [
 export const setAntUndernameOptions = [
   ...setAntBaseNameOptions,
   optionMap.undername,
+];
+
+export const upgradeAntOptions = [
+  optionMap.processId,
+  optionMap.names,
+  optionMap.reassignAffiliatedNames,
+  ...writeActionOptions,
 ];
 
 export const transferRecordOwnershipOptions = [